refactor and simplify broadcast group configuration
* removed all implementation classes from the api and put them inside
  API classes
* JGroupsBroadcastConfiguration can takes either a jgroups fileName
  string or a JChannel object
